fix: Use BigInteger for user ID to support large Telegram IDs

+++ bot/src/infra/db/models/deck.py
@@ -1,6 +1,6 @@
 from typing import TYPE_CHECKING
 
-from sqlalchemy import ForeignKey, Integer, String
+from sqlalchemy import BigInteger, ForeignKey, String
 from sqlalchemy.orm import Mapped, mapped_column, relationship
 
 from .base import Base
@@ -13,5 +13,5 @@ if TYPE_CHECKING:
 class Deck(Base, AuditMixin):
     __tablename__ = "decks"
     name: Mapped[str] = mapped_column(String(255), nullable=False)
-    user_id: Mapped[int] = mapped_column(Integer, ForeignKey("users.id"), nullable=False)
+    user_id: Mapped[int] = mapped_column(BigInteger, ForeignKey("users.id"), nullable=False)
     cards: Mapped[list[Card]] = relationship("Card", cascade="all, delete-orphan")
