+++ bot/src/infra/db/models/user.py
@@ -1,6 +1,6 @@
 from typing import TYPE_CHECKING
 
-from sqlalchemy import Boolean, Integer, String
+from sqlalchemy import BigInteger, Boolean, Integer, String
 from sqlalchemy.orm import Mapped, mapped_column, relationship
 
 from .base import Base
@@ -11,7 +11,7 @@ if TYPE_CHECKING:
 
 class User(Base):
     __tablename__ = "users"
-    id: Mapped[int] = mapped_column(Integer, primary_key=True)  # tg user id
+    id: Mapped[int] = mapped_column(BigInteger, primary_key=True)  # tg user id
     name: Mapped[str] = mapped_column(String(255), nullable=False)
     onboarding_completed: Mapped[bool] = mapped_column(
         Boolean, default=False, server_default="false"
